pyproject.toml
[tool.poetry]
name = "food-app"
version = "0.1.0"
description = ""
authors = ["Maria.Kowalska <maria.kowalska@itechart-group.com>"]
readme = "README.md"
packages = [{include = "food_app"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2 = "^2.9.5"
dj-database-url = "^1.2.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
